Change Python output strategy.
Instead of creating a stepping stone error (opensees.msg), we can pass
things directly to Python's stderr using PySys_FormatStderr. This output
can be caught and redirected by things expecting it (e.g. Python's
contextlib.redirect_stderr). Additionally, the last message printed is
saved, so when a real error occurs, it has that information along with
it.

Finally, I've renamed opensees.error to opensees.OpenSeesError and made
it available from the actual module. This allows it to be caught by the
calling Python script.