Take more number classes as input
Currently, the Python module needs int and float inputs to be the
precise class (or subclass of) the input requested. We can skip the
up-front check for this, and have the Python internals call the
appropriate conversion method if it can. If it can't convert the number,
it sets the error state and we can just pass this up the chain as usual.