Fixing bug in DKG family of shell elements - wrong Bbend sign
DKG  family of shell elements (ShellDKGQ, ShellDKGT, ShellNLDKGQ, ShellNLDKGT and ShellDKGQThermal) show a subtle behaviour in bending only visible when used with LayeredShellSection.
They compute the curvature with a sign which is inverted with respect to what the layered shell expects. However they post-multiply the bending B-matrix so when they compute B^T*S the sign of the internal forces is correct.
However this does not give correct results when used with composite sections, with nonlinear material that behave differently in tension and compression