Attention!
This commit adds new staged construction functionality that makes it easier to do this type of analysis in the geotechnical setting and
also works in parallel. The idea is to create all elements and nodes at the start, and only activate/deactivate elements accordingly. Nodal
activation/deactivation is handled by adding 1 to the diagonal of the stiffness matrix and keeping zeros on the RHS, this ensures these
dofs are not updated.

It is also important that elements implement activate functionality by overloading the appropriate functions.

Finally, care should be taken to consider initial displacements in the elements.

More details:

* Changes to LoadControl and Newmark classes are there so that I can subclass those into their staged versions
* Added functionality to Element and FE_Element class (this is all handled in the FE_Element)
* New integrators that support the staged construction system.
* New class tags for the new integrators