Added ndm and ndf
Utility commands for getting the input ndm and input ndf, as well as the nodal ndm and ndf. Syntax:
ndm <$nodeTag>
ndf <$nodeTag>
If the nodeTag argument is provided, it will return the nodal dimensions/degrees of freedom, or throw an error if the node does not exist.
If the nodeTag argument is not provided, it will return the input ndm/ndf, or the blank string if the model builder is undefined.