Create MaterialTest.cpp
Created MaterialTest
Design intent is to be a combination of uniaxial material, ND material, and section testers. Main commands:
testUniaxialMaterial
testNDMaterial
testSection
Each of the main test commands will define the test type (1 for uniaxial, 2 for ND, 3 for section), and set the static object to the corresponding material/section.
The other commands will each have a switch based on test type (0, 1, 2, and 3). 0 is the default, and will throw an error
The other commands are taken from the uniaxial material test commands in the python interpreter, with the exception that setTrialStrain, commitStrain, and getResponse will be added.