Update QuadPatch.cpp
Hi
This update is for QuadPath to always get Positive Areas and modify clockwise definition and make them counter clockwise. Because I'm not perfect in C++ please check if did I use correct #include (lines 35 - 37) and also is definition of vertexCoordtmp is correct?
The algorithm check in DevC++ by various samples.