Disable OTHER/CMakeList to see what breaks
The CMakeList.txt in the top of the OTHER directory appears to be unnecessary.

Commented entire file to see what breaks. The project referenced in this CMake List does not appear to be used anywhere else in the main project. If this breaks something (not related to building the core OpenSees program), then this should be refactored.