Update CMakeLists.txt
The current method hard-coded directories and sent them to `include_directories(...)` AND THEN used a recursive glob method to collect the same directories and sent them (again) to `include_directories(...)`. This caused the linker command line to be massively long and caused `command line too long` errors .

Additionally, using `CMakeList.txt` as the search term rather than `*.h` is the preferred method. This ensures that only directories that can actually be built/included with CMake are included.... No CMakeList.txt, no CMake build.

`  #${OPS_BUNDLED_DIR}/CMakeLists.txt` has been commented, since this line is likely the root of a lot of compile issues that occur. This line was adding all of the /OTHER directories to the `include_directories(...)`  which leads to conflicts during linking when there are multiple versions of the same package and/or user-supplied external libraries