[FIX] Update PVD recorder to handle one element request per "eleResponse" input
This change allows the user to request multiple element outputs from a single PVD recorder. Each time the 'eleResponse' keyword is identified it is treated independently of previous requests. The previous implementation resulted in all requests following the keyword to be concatenated and only the first request to actually be output in the PVD recorder with an incorrect name.